Add spellcheck and trailing whitespace checks.
file/spelling_error uses whoosh to check for spelling errors in
comments and docstrings. It will also detect use of stale symbols
in comments and docstirngs. User-defined terms can be sourced
from a file called DICTIONARY in the project directory.

Pass --spellcheck-cache with a directory name to specify a
directory to cache the spellcheck graph between invocations.
Generating the spellcheck graph can take a very long time, so this
will speed up execution substantially.

Pass --log-technical-terms-to to note any symbols seen in the code
linted to the file specified. The result will be stored as the
logical union of the old file contents and the newly seen terms.

Pass --block-regexes to cause any text matching the specified
regex not to be spellchecked or checked against the list of used
symbols at all. This is particularly useful for inline suppressions
which rely on code-like syntax which might otherwise be detected
as a symbol.

This change also includes the spellcheck-linter tool, which will
run a spellcheck on the entire file. Pass --technical-terms and
a generated file containing techncial terms to also check for
usage of symbols which don't exist in surrounding code. This
is particularly useful for checking documentation.