spelling: Add new _find_spellcheckable_chunks function
This does not use regex scanning at all and instead is O(N), going
ever each character at a time and maintaining some state. This is
a lot more maintainable, even if the implementation is
slightly more complex.

This fixes a nasty bug where a single quotation mark inside a
comment was being counted as a quotation start point
and causing stuff in comments to be counted
as "techincal" and worse, stuff outside of comments to be
counted as spellcheckable